Use SpinSleep instead of regular sleep to avoid pegging the CPU
The default Windows sleep resolution is about 15 miliseconds, so if you
sleep for any amount of time, it will either do nothing or essentially
drop a whole frame for you. This feels awful to play and also causes
desync issues over time. Instead we use SpinSleep which attempts to
sleep for a short amount of time and then just spins until the desired
duration is achieved. This should hopefully allow less CPU to be used
(ie: the game will not run at 400 FPS for no reason) while not causing
awful sub-45 jerky FPS gameplay.