Switch kira for audio playback.
The reason to use kira over ggez's (really rodio's) audio capabilities
is that rodio skips ahead in songs by consuming samples from the audio
in `frame_len` chunks until the desired duration is skipped. This is
terrible! It means that, in practice, you skip 50 seconds of audio by
consuming 2 million samples, 1000 samples at a time. This causes a huge
lag spike (and wastes a bunch of CPU for no reason), which explains why
there's desync if you skip to late in a file. Kira avoid this problem,
so we can skip ahead freely. This gives a nice boost in performance and
prevents desync. Hurray!